powerpc: dt_cpu_ftrs: Set current thread fscr bits
Setting the FSCR bit directly in the SPR only sets it during the
initial boot and early init of the kernel but not for init or any
subsequent kthreads as the thread_struct for those is copied from
current which doesn't reflect any changes made during cpu feature
detection. This patch ensures the current thread state is updated to
match.

Signed-off-by: Alistair Popple <alistair@popple.id.au>